Decoupled OpenIDConnect from authentication
