+++ alchemy/admin_server/__init__.py
@@ -24,14 +24,17 @@ if env.bool("USE_CLOUD_LOGGING", default=False):
 def create_app(test_config=None):
     # create and configure the app
     app = Flask(__name__, instance_relative_config=True)
-    app.config.from_mapping(SECRET_KEY="athena_todo_change_this_in_prod")
-
     if test_config is None:
         # load the instance config, if it exists, when not testing
         app.config.from_object(DevelopmentConfig)
     else:
         # load the test config if passed in
         app.config.from_object(test_config)
+
+    app.config.update({
+        'SECRET_KEY': 'athena_todo_change_this_in_prod',
+        'OKTA_BACKEND': 'alchemy.shared.okta.openid_connect',
+    })
 
     # ensure the instance folder exists
     try:
